Load Interface and Firewall drivers by name
Neutron uses stevedore to load these drivers by name, and when that
fails it looks them up by classname.  A change in stevedore in newton
added a warn_on_missing_entrypoint to the NamedExtensionManager with a
default to True.  Neutron does not set this parameter, so by default it
will log a Warning if you load the driver by class name.

By updating the config to load them by the alias, those warnings no
longer happen, and less code needs to execute to find the class.