Remove deprecated option external_network_bridge
This option was changed to default to "" in newton, and marked as
deprecated for removal. Since this isn't configurable, leaving it to
default to "" will avoid warnings on service startup.

Updated the Help text for this setting on newton to match that of a
newton release version of neutron-l3-agent.